+++ android/src/main/kotlin/tv/mta/flutter_playout/video/PlayerLayout.java
@@ -606,6 +606,31 @@ public class PlayerLayout extends PlayerView implements FlutterAVPlayer, EventCh
         } catch (Exception e) { /* ignore */ }
     }
 
+    /**
+     * set text track language for player - language must be one of available in HLS manifest
+     * currently playing or from the array of text tracks passed to player
+     *
+     * @param arguments
+     */
+    public void setPreferredTextLanguage(Object arguments) {
+        try {
+
+            java.util.HashMap<String, String> args = (java.util.HashMap<String, String>) arguments;
+
+            String languageCode = args.get("code");
+
+            this.preferredTextLanguage = languageCode;
+
+            if (mPlayerView != null && trackSelector != null && mPlayerView.isPlaying()) {
+
+                trackSelector.setParameters(
+                        trackSelector.buildUponParameters()
+                                .setPreferredTextLanguage(languageCode));
+            }
+
+        } catch (Exception e) { /* ignore */ }
+    }
+
     public void seekTo(Object arguments) {
         try {